+++ lib/groupdocs/document/page.rb
@@ -1,0 +1,13 @@
+#Added in release 2.1.0
+module GroupDocs
+  class Document::Page < Api::Entity
+
+    # @attr [Integer] Id
+    attr_accessor :Id
+    # @attr [Integer] Width
+    attr_accessor :Width
+    # @attr [Integer] Height
+    attr_accessor :Height
+
+  end #Document::Page
+end #GroupDocs